backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.3.0",
    "channels-redis>=4.3.0",
    "daphne>=4.1.0",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=24.1.1",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "channels[daphne]>=4.3.2",
    "django-debug-toolbar>=6.2.0",
    "model-bakery>=1.23.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.14",
]

[tool.pyright]
extraPaths = ["./"]
typeCheckingMode = "basic"

[tool.ruff.format]
quote-style = "single"
